alx_backend_caching_property_listings/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-redis==6.0.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
psycopg2-binary==2.9.10
redis==6.2.0
sqlparse==0.5.3
